requirements_agent.txt
# Windows 에이전트용
python-socketio[client]>=4.6.0,<5.0
pyserial>=3.5
